Move types to @learncard/types
