+++ services/learn-card-network/brain-service/src/accesslayer/inbox-credential/update.ts
@@ -1,6 +1,6 @@
 import { QueryBuilder, BindParam } from 'neogma';
 import { InboxCredential } from '@models';
-import { InboxCredentialType } from 'types/inbox-credential';
+import { InboxCredentialType } from '@learncard/types';
 import { flattenObject, inflateObject } from '@helpers/objects.helpers';
 
 export const updateInboxCredential = async (
